backend/pyproject.toml
[tool.poetry]
name = "cafeteria-backend"
version = "0.1.0"
description = "Backend FastAPI para sistema de cafetería"
authors = ["Equipo Cafetería <dev@cafeteria.local>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlalchemy = "^2.0.25"
alembic = "^1.12.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.2"
ruff = "^0.2.0"

[tool.poetry.scripts]
app = "app.main:run"

[build-system]
requires = ["poetry-core>=1.6.1"]
build-backend = "poetry.core.masonry.api"
